backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0

sqlalchemy[asyncio]>=2.0.30
aiosqlite>=0.20.0
alembic>=1.13.0

pydantic>=2.7.0
pydantic-settings>=2.2.0

python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
argon2-cffi>=23.1.0
python-multipart>=0.0.9
email-validator>=2.1.1

redis>=5.0.4
celery>=5.4.0

numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0

torch>=2.0.0
transformers>=4.38.0
sentence-transformers>=2.2.2

stable-baselines3>=2.2.1
gymnasium>=0.29.1

accelerate>=0.28.0
